+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ucentral-client",
-  "version": "0.9.14",
+  "version": "0.9.15",
   "dependencies": {
     "@coreui/coreui": "^3.4.0",
     "@coreui/icons": "^2.0.1",
@@ -21,7 +21,7 @@
     "react-router-dom": "^5.2.0",
     "react-widgets": "^5.1.1",
     "sass": "^1.35.1",
-    "ucentral-libs": "^0.8.7",
+    "ucentral-libs": "^0.8.8",
     "uuid": "^8.3.2"
   },
   "main": "index.js",
