--- src/layout/Sidebar/index.module.scss
@@ -1,9 +1,0 @@
-.sidebarImgFull {
-  height: 100px;
-  width: 230px;
-}
-
-.sidebarImgMinimized {
-  height: 75px;
-  width: 75px;
-}
