+++ src/layout/index.js
@@ -3,23 +3,29 @@ import { useTranslation } from 'react-i18next';
 import { logout } from 'utils/authHelper';
 import routes from 'routes';
 import { useAuth } from 'contexts/AuthProvider';
-import { Header } from 'ucentral-libs';
-import Sidebar from './Sidebar';
-import TheContent from './Content';
-import TheFooter from './Footer';
+import { Header, Sidebar, Footer, PageContainer } from 'ucentral-libs';
 
 const TheLayout = () => {
   const [showSidebar, setShowSidebar] = useState('responsive');
   const { endpoints, currentToken } = useAuth();
   const { t, i18n } = useTranslation();
 
+  const navigation = [
+    {
+      _tag: 'CSidebarNavItem',
+      name: t('common.device_list'),
+      to: '/devices',
+      icon: 'cilNotes',
+    },
+  ];
+
   return (
     <div className="c-app c-default-layout">
       <Sidebar
         showSidebar={showSidebar}
         setShowSidebar={setShowSidebar}
-        t={t}
-        logo="assets/OpenWiFi_LogoLockup_DarkGreyColour.svg"
+        logo="assets/OpenWiFi_LogoLockup_WhiteColour.svg"
+        options={navigation}
       />
       <div className="c-wrapper">
         <Header
@@ -33,9 +39,9 @@ const TheLayout = () => {
           endpoints={endpoints}
         />
         <div className="c-body">
-          <TheContent />
+          <PageContainer t={t} routes={routes} />
         </div>
-        <TheFooter />
+        <Footer t={t} version="0.9.15" />
       </div>
     </div>
   );
